Move: colorStringToHex.js를 fixture/common으로
