+++ fixture/common/colorStringToHex.js
